feat: implement fragment for the rv
